+++ WalkLive/WalkLive/Cell/FriendTripTableViewCell.swift
@@ -14,7 +14,9 @@ class FriendTripTableViewCell: UITableViewCell {
     @IBOutlet weak var sourceLabel: UILabel!
     @IBOutlet weak var destinationLabel: UILabel!
     @IBOutlet weak var timeSpentLabel: UILabel!
-    
+    @IBOutlet weak var phoneButton: UIButton!
+    @IBOutlet weak var messageButton: UIButton!
+    @IBOutlet weak var mapButton: UIButton!
     
     override func awakeFromNib() {
         super.awakeFromNib()