+++ WalkLive/WalkLive/VC/UserVC.swift
@@ -7,25 +7,64 @@
 //
 
 import UIKit
+import MessageUI
 
-class UserVC: UIViewController {
+
+class UserVC: UIViewController, UITableViewDelegate, UITableViewDataSource, MFMessageComposeViewControllerDelegate {
+    
     @IBOutlet weak var usernameLabel: UILabel!
     
     @IBOutlet weak var userContactLabel: UILabel!
     
     @IBOutlet weak var emergencyContactLabel: UILabel!
     @IBOutlet weak var friendRequestButton: UIButton!
+    
+    @IBOutlet weak var friendTripTable: UITableView!
+    
+    
     var receivedFriendRequests: [FriendRequest]!
+    var friendsTrip: [TimePoint]?
     
-    
+    var timeInterval = TimeInterval(20)
+    var tripUpdater: TripUpdater!
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        friendTripTable.delegate = self
+        friendTripTable.dataSource = self
+        self.updateFriendTrips()
         // Do any additional setup after loading the view.
     }
     
     override func viewWillAppear(_ animated: Bool) {
         self.setUserVCInfo()
+        self.initializeUpdates()
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        print("closed successfully")
+        self.endTimers()
+        
+    }
+    
+    private func initializeUpdates(){
+        self.tripUpdater = TripUpdater(tripTableDelegate: self)
+        self.tripUpdater.startTimer(timeInterval: self.timeInterval)
+    }
+    
+    private func endTimers(){
+        self.tripUpdater.endTimer()
+    }
+    
+    func updateFriendTrips(){
+        backEndClient.getAllTrip(success: { (timePoints) in
+            OperationQueue.main.addOperation {
+                self.friendsTrip = timePoints
+                self.friendTripTable.reloadData()
+            }
+        }) { (error) in
+            print("failed to get all friend trips")
+        }
     }
     
     
@@ -50,6 +89,33 @@ class UserVC: UIViewController {
         self.performSegue(withIdentifier: "settingSegue", sender: self)
     }
     
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        if self.friendsTrip != nil {
+            print("reached here")
+            return self.friendsTrip!.count
+        } else {
+            print("reached here for 0 ")
+            return 0
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = friendTripTable.dequeueReusableCell(withIdentifier: "FriendTripTableViewCell", for: indexPath) as! FriendTripTableViewCell
+        let friendTrip = self.friendsTrip![indexPath.row]
+        cell.friendNameLabel.text = friendTrip.username
+        cell.destinationLabel.text = "To: " + friendTrip.destination!
+        cell.sourceLabel.text = "At: (" + String(format: "%.3f", (friendTrip.curLat)!) + ", " + String(format: "%.3f", (friendTrip.curLong)!) + ")"
+        cell.timeSpentLabel.text = friendTrip.timeSpent
+        cell.phoneButton.tag = indexPath.row
+        cell.messageButton.tag = indexPath.row
+        cell.mapButton.tag = indexPath.row
+       // cell.messageButton.tag = indexPath.row
+        cell.selectionStyle = .none
+        return cell
+    }
+    
+    
     func updateFriendRequest(){
         backEndClient.getReceivedFriendRequests(success: { (receivedFriendRequests) in
             OperationQueue.main.addOperation {
@@ -72,6 +138,39 @@ class UserVC: UIViewController {
         }
     }
     
+    @IBAction func onMessageButton(_ sender: UIButton) {
+        let phoneNum = friendsTrip![sender.tag].emergencyNum
+        if (MFMessageComposeViewController.canSendText()) {
+            let controller = MFMessageComposeViewController()
+            controller.body = "Hey! Where are you?" + "From: \(currentUserInfo.username!)"
+            controller.recipients = [phoneNum!]
+            controller.messageComposeDelegate = self
+            self.present(controller, animated: true, completion: nil)
+        }
+    }
+    
+
+    @IBAction func onPhoneButton(_ sender: UIButton) {
+        let phoneNum = friendsTrip![sender.tag].emergencyNum
+        //Does not work in Simulator
+        if let url = URL(string: "tel://\(phoneNum!)"), UIApplication.shared.canOpenURL(url) {
+            if #available(iOS 10, *) {
+                UIApplication.shared.open(url)
+            } else {
+                UIApplication.shared.openURL(url)
+            }
+        }
+    }
+    
+    
+    func messageComposeViewController(_ controller: MFMessageComposeViewController, didFinishWith result: MessageComposeResult) {
+        //... handle sms screen actions
+        self.dismiss(animated: true, completion: nil)
+    }
+    
+    
+    
+    
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
@@ -84,7 +183,7 @@ class UserVC: UIViewController {
      // In a storyboard-based application, you will often want to do a little preparation before navigation
      override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         let backItem = UIBarButtonItem()
-        backItem.title = "Back"
+        backItem.title = currentUserInfo.username
         backItem.tintColor = primaryColor
         navigationItem.backBarButtonItem = backItem
         if (segue.identifier == "friendRequestSegue"){
@@ -92,10 +191,29 @@ class UserVC: UIViewController {
             vc.receivedFriendRequests = self.receivedFriendRequests
         } else if (segue.identifier == "settingSegue"){
         } else if (segue.identifier == "tripRequestSegue"){
+        } else if (segue.identifier == "friendTrackerSegue"){
+            let vc = segue.destination as! FriendTrackVC
+            let button = sender as! UIButton
+            vc.currentFriendTrip = friendsTrip?[button.tag]
+            vc.friendTrips = self.friendsTrip
         }
      // Get the new view controller using segue.destinationViewController.
      // Pass the selected object to the new view controller.
      }
     
+}
+
+extension UserVC: TripTableUpdateDelegate{
+    func updateTable() {
+        backEndClient.getAllTrip(success: { (friendTrips) in
+            OperationQueue.main.addOperation {
+                self.friendsTrip = friendTrips
+                self.friendTripTable.reloadData()
+                print("reloaded table!")
+            }
+        }) { (error) in
+            print("failed to update")
+        }
+    }
 }